Fix last commit in detail_path_manager.h

+++ src/xrGame/detail_path_manager.h
@@ -79,7 +79,7 @@ public:
 
 protected:
     VELOCITIES m_movement_params;
-    u64 m_current_travel_point;
+    size_t m_current_travel_point;
 
 private:
     enum EDirectionType
